% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-add.r
\name{fmrutils.add}
\alias{fmrutils.add}
\title{Simple Add Function}
\usage{
fmrutils.add(a, b)
}
\arguments{
\item{a}{first integeger}

\item{b}{second integer}
}
\value{
output integer value
}
\description{
This function adds two integers
}
\examples{
  fmrutils.add(10, 20)
}
