% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-add-range.r
\name{fmrutils.addrange}
\alias{fmrutils.addrange}
\title{Simple Add Function}
\usage{
fmrutils.addrange(from = 1, to = 10)
}
\arguments{
\item{from}{start integeger}

\item{to}{second integer}
}
\value{
output summed integer value
}
\description{
This function adds all values from a range starting from an integer to target integer
}
\examples{
  fmrutils.addrange(10, 20)
  fmrutils.addrange(100)
  fmrutils.addrange()
  fmrutils.addrange(to=100, from= 10)
}
